FEAT: Phase 2 model-first Parameter contract and kind-driven resolver
Move the declarative Parameter contract to pyrit/models/parameter.py as the
canonical home, add ComponentRegistryKind / RegistryReference / ParameterDestination,
and re-key the registry resolver off registry kind instead of inline issubclass checks.
pyrit/common/parameter.py becomes a compatibility re-export.
